# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/test2": {
            "get": {
                "tags": [
                    "首页"
                ],
                "summary": "测试",
                "responses": {
                    "200": {
                        "description": "welcome",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/createUser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "新增用户",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "user_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "passWord",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "名称",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "phone",
                        "name": "phone",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/deleteUser": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "userName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getUserByPhone": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "根据手机号获取用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getUserByUEmail": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "根据邮箱获取用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getUserByUserName": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "根据userName获取用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号",
                        "name": "userName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getUserList": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "所有用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "user-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "user-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/loginUser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "登录用户",
                "parameters": [
                    {
                        "description": "账号",
                        "name": "userName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "passWord",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/signOut": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "登录登出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "user-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "user-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/updateUser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "修改用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号",
                        "name": "userName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "passWord",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": \"200\", \"message\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}